# SchoolDiarySystem.WebClient/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SchoolDiarySystem.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Student": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Получает всех студентов.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Student"
        ],
        "summary": "Создает нового студента.",
        "requestBody": {
          "description": "Данные для создания нового студента.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Возвращает созданного студента.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGetDto"
                }
              }
            }
          },
          "400": {
            "description": "Если данные студента некорректны.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Student/{id}": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Получает студента по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор студента.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Возвращает студента.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGetDto"
                }
              }
            }
          },
          "404": {
            "description": "Если студент не найден.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Student"
        ],
        "summary": "Обновляет существующего студента.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор студента для обновления.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Обновленные данные студента.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Возвращает обновленного студента.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGetDto"
                }
              }
            }
          },
          "400": {
            "description": "Если данные студента некорректны.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Если студент не найден.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Student"
        ],
        "summary": "Удаляет студента по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор студента для удаления.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Возвращает сообщение об успешном удалении."
          },
          "404": {
            "description": "Если студент не найден.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "StudentGetDto": {
        "required": [
          "birthDate",
          "classId",
          "fullName",
          "passport",
          "studentId"
        ],
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "passport": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "description": "ФИО ученика.",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения ученика.",
            "format": "date-time"
          },
          "classId": {
            "type": "integer",
            "description": "Идентификатор класса ученика.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "StudentPostDto": {
        "required": [
          "birthDate",
          "classId",
          "fullName",
          "passport"
        ],
        "type": "object",
        "properties": {
          "passport": {
            "type": "string",
            "description": "Паспортные данные ученика.",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "description": "ФИО ученика.",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения ученика.",
            "format": "date-time"
          },
          "classId": {
            "type": "integer",
            "description": "Идентификатор класса ученика.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}